remove unnecessary cast

+++ include/hri_safe_remote_control_system/VscProcess.h
@@ -52,7 +52,7 @@ struct ErrorCounterType
  */
 const unsigned int VSC_INTERFACE_RATE = 50; /* 50 Hz */
 const unsigned int VSC_HEARTBEAT_RATE = 20; /* 20 Hz */
-const unsigned int VSC_PAUSE_SETTINGS_PERIOD_S = 5; /* 0.2 Hz */
+const double VSC_PAUSE_SETTINGS_PERIOD_S = 5; /* 0.2 Hz */
 
 class VscProcess
 {
